+++ src/main/java/ht/henrique/mazebank/model/fetch/FetchUserResponse.java
@@ -4,10 +4,13 @@ import ht.henrique.mazebank.model.Response;
 import jakarta.persistence.Column;
 import lombok.Data;
 
+import java.math.BigDecimal;
+import java.text.DecimalFormat;
+
 @Data
 public class FetchUserResponse implements Response {
     private Integer uid;
     private String userName;
-    private Integer userBalance;
+    private BigDecimal userBalance;
     private String userEmail;
 }
